style(lint): clear swiftlint violations across files, models, services, and views
